requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
openai==1.35.14